requirements.txt
streamlit
streamlit-drawable-canvas
tensorflow
numpy
opencv-python
Pillow
